requirements.txt
python-pptx==1.0.2
loguru==0.7.2
gradio==5.4.0
langchain==0.3.6
langchain-openai==0.2.4
langchain_community==0.3.4
python-dotenv==1.0.1
requests==2.32.3
